+++ client/include/pm3.h
@@ -16,10 +16,13 @@
 #ifndef LIBPM3_H
 #define LIBPM3_H
 
+#include <stdbool.h>
+
 typedef struct pm3_device pm3;
 
 pm3 *pm3_open(const char *port);
-int pm3_console(pm3 *dev, const char *cmd);
+int pm3_console(pm3 *dev, const char *cmd, bool passthru);
+const char *pm3_grabbed_output_get(pm3 *dev);
 const char *pm3_name_get(pm3 *dev);
 void pm3_close(pm3 *dev);
 pm3 *pm3_get_current_dev(void);